+++ src/openlayers/overlay/graphic/index.js
@@ -1,3 +1,6 @@
 export {Graphic as OverlayGraphic} from './Graphic';
 export {CloverShape} from './CloverShape';
 export {HitCloverShape} from './HitCloverShape';
+
+export {GraphicCanvasRenderer} from './CanvasRenderer';
+export {GraphicWebGLRenderer} from './WebGLRenderer';