+++ test/karma.conf.js
@@ -55,7 +55,7 @@ module.exports = function (config) {
             /***openlayers的源码***/
             {pattern: '../node_modules/openlayers/dist/ol-debug.css', include: false},
             {pattern: '../src/openlayers/**/**/*.css', include: false},
-            '../src/openlayers/**/!(index|MapboxStyles).js',
+            '../src/openlayers/**/!(Graphic|WebGLRenderer|index|MapboxStyles).js',
             /**测试文件**/
             './test-main-openlayers.js',
 
@@ -91,7 +91,7 @@ module.exports = function (config) {
             './test-main-leaflet.js': ['browserify'],
 
             '../node_modules/openlayers/dist/ol-debug.js': ['browserify'],
-            '../src/openlayers/**/!(index).js': ['browserify'],
+            '../src/openlayers/**/!(Graphic|WebGLRenderer|index).js': ['browserify'],
             './openlayers/**/*Spec.js': ['browserify'],
             './test-main-openlayers.js': ['browserify'],